Revert "He cambiado todos los atributos double a decimal"
This reverts commit 3c14e62643242ee4781f99a54d90c7274c6dc484.

+++ src/AppForSEII2526.API/Models/RentalItem.cs
@@ -14,7 +14,7 @@
             RentalId = rental.Id;
         }
 
-        public RentalItem(int carId, Rental rental, decimal priceForRenting)
+        public RentalItem(int carId, Rental rental, double priceForRenting)
         {
             CarId = carId;
             Rental = rental;
@@ -30,6 +30,6 @@
 
         [Range(1, int.MaxValue, ErrorMessage = "La cantidad debe ser mayor que 1")]
         public int Quantity { get; set; }
-        public decimal PriceForRenting { get; set; }
+        public double PriceForRenting { get; set; }
     }
 }
